third_party/WebKit/LayoutTests/platform/win/external/wpt/html/semantics/forms/textfieldselection/selection-not-application-textarea-expected.txt

This is a testharness.js-based test.
FAIL text field selection for the input textareaassert_equals: initial selectionDirection expected "none" but got "forward"
Harness: the test ran to completion.
